Develop

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 
@@ -8,7 +8,7 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="This field cannot be left blank!")
     parser.add_argument('store_id', type=float, required=True, help="Every item needs a store_id!")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
 
@@ -16,6 +16,7 @@ class Item(Resource):
             return item.json()
         return {"message": "Item not found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists".format(name)}, 400
@@ -45,7 +46,12 @@ class Item(Resource):
 
         return item.json()
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privileges required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -54,5 +60,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {'items': items}
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }
